AddProperty : Ajout photo à partir de la galery

+++ app/src/main/java/com/openclassrooms/realestatemanager/ui/detail_property/DetailPropertyFragment.kt
@@ -40,7 +40,7 @@ class DetailPropertyFragment : Fragment() {
 
         binding = FragmentDetailPropertyBinding.inflate(inflater, container, false)
 
-        val idProperty = arguments!!.getLong("idProperty")
+        val idProperty = arguments!!.getInt("idProperty")
 
         viewModel.getPropertyById(idProperty).observe(this) { property ->
             if (property!=null) {
